src/main/java/com/example/demo/controller/AppInfoController.java
package com.example.demo.controller;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.HashMap;
import java.util.Map;

@RestController
@RequestMapping("/app")
public class AppInfoController {

    @Value("${spring.application.name}")
    private String appName;

    @Value("${server.port}")
    private String port;

    @GetMapping("/info")
    public Map<String, String> getAppInfo() {
        Map<String, String> info = new HashMap<>();
        info.put("appName", appName);
        info.put("port", port);
        return info;
    }

    @GetMapping("/name")
    public String getName() {
        return appName;
    }

    @GetMapping("/port")
    public String getPort() {
        return port;
    }
}
